blog page showing the issue
This commit includes ony the Blog page of my website containing the issue I am facing. The issue is that there are images with the data-pin-nopin attribute used to prevent these images showing up when the user clicks on the pinterest share button. For uknown reason, the unwanted images show up. I'm still looking for ways to fix this long standing issue.